pyproject.toml
[tool.poetry]
name = "flame"
version = "0.1.0"
description = ""
authors = ["atlasatlas <68698872+westoleaboat@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
django = "4.2.4"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
django-registration = "^3.4"
psycopg2 = "^2.9.9"
oauth2-provider = "^0.0"
django-oauth-toolkit = "^2.3.0"
python-dotenv = "^1.0.1"
channels = "^4.0.0"
django-taggit = "^5.0.1"
django-markdownx = "^4.0.7"
django-markdown-deux = "^1.0.6"
markdown = "^3.5.2"
markdown2 = "^2.4.12"
markdownify = "^0.11.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
